+++ EMS/core-bundle/src/Controller/ContentManagement/ContentTypeController.php
@@ -6,13 +6,13 @@ use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
 use EMS\CommonBundle\Helper\EmsFields;
+use EMS\CoreBundle\Core\Form\FieldTypeManager;
 use EMS\CoreBundle\Entity\ContentType;
 use EMS\CoreBundle\Entity\Environment;
 use EMS\CoreBundle\Entity\FieldType;
 use EMS\CoreBundle\Entity\Form\ContentTypeJsonUpdate;
 use EMS\CoreBundle\Entity\Form\EditFieldType;
 use EMS\CoreBundle\Exception\ElasticmsException;
-use EMS\CoreBundle\Form\DataField\DataFieldType;
 use EMS\CoreBundle\Form\DataField\SubfieldType;
 use EMS\CoreBundle\Form\Field\IconTextType;
 use EMS\CoreBundle\Form\Field\SubmitEmsType;
@@ -37,7 +37,6 @@ use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Form;
 use Symfony\Component\Form\FormError;
-use Symfony\Component\Form\FormRegistryInterface;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -47,17 +46,22 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class ContentTypeController extends AbstractController
 {
-    public function __construct(private readonly LoggerInterface $logger, private readonly ContentTypeService $contentTypeService, private readonly Mapping $mappingService, private readonly FormRegistryInterface $formRegistry)
+    public function __construct(
+        private readonly LoggerInterface $logger,
+        private readonly ContentTypeService $contentTypeService,
+        private readonly Mapping $mappingService,
+        private readonly FieldTypeManager $fieldTypeManager)
     {
     }
 
+    /**
+     * @deprecated
+     */
     public static function isValidName(string $name): bool
     {
-        if (\in_array($name, [Mapping::HASH_FIELD, Mapping::SIGNATURE_FIELD, Mapping::FINALIZED_BY_FIELD, Mapping::FINALIZATION_DATETIME_FIELD])) {
-            return false;
-        }
+        @\trigger_error('Deprecated isValidName function, please use the FieldTypeManager::isValidName function', E_USER_DEPRECATED);
 
-        return \preg_match('/^[a-z][a-z0-9\-_]*$/i', $name) && \strlen($name) <= 100;
+        return FieldTypeManager::isValidName($name);
     }
 
     public function updateFromJsonAction(ContentType $contentType, Request $request): Response
@@ -362,63 +366,6 @@ class ContentTypeController extends AbstractController
         ]);
     }
 
-    /**
-     * Try to find (recursively) if there is a new field to add to the content type.
-     *
-     * @param array<mixed> $formArray
-     *
-     * @throws ElasticmsException
-     */
-    private function addNewField(array $formArray, FieldType $fieldType): bool|string
-    {
-        if (\array_key_exists('add', $formArray)) {
-            if (isset($formArray['ems:internal:add:field:name'])
-                && 0 != \strcmp((string) $formArray['ems:internal:add:field:name'], '')
-                && isset($formArray['ems:internal:add:field:class'])
-                && 0 != \strcmp((string) $formArray['ems:internal:add:field:class'], '')) {
-                if (static::isValidName($formArray['ems:internal:add:field:name'])) {
-                    $fieldTypeNameOrServiceName = $formArray['ems:internal:add:field:class'];
-                    $fieldName = $formArray['ems:internal:add:field:name'];
-                    /** @var DataFieldType $dataFieldType */
-                    $dataFieldType = $this->getDataFieldType($fieldTypeNameOrServiceName);
-                    $child = new FieldType();
-                    $child->setName($fieldName);
-                    $child->setType($fieldTypeNameOrServiceName);
-                    $child->setParent($fieldType);
-                    $child->setOptions($dataFieldType->getDefaultOptions($fieldName));
-                    $fieldType->addChild($child);
-                    $this->logger->notice('log.contenttype.field.added', [
-                        'field_name' => $fieldName,
-                        EmsFields::LOG_OPERATION_FIELD => EmsFields::LOG_OPERATION_CREATE,
-                    ]);
-
-                    return '_ems_'.$child->getName().'_modal_options';
-                } else {
-                    $this->logger->error('log.contenttype.field.name_not_valid', [
-                        'field_format' => '/[a-z][a-z0-9_-]*/ !'.Mapping::HASH_FIELD.' !'.Mapping::HASH_FIELD,
-                    ]);
-                }
-            } else {
-                $this->logger->error('log.contenttype.field.name_mandatory', [
-                ]);
-            }
-
-            return true;
-        } else {
-            /** @var FieldType $child */
-            foreach ($fieldType->getChildren() as $child) {
-                if (!$child->getDeleted()) {
-                    $out = $this->addNewField($formArray['ems_'.$child->getName()], $child);
-                    if (false !== $out) {
-                        return '_ems_'.$child->getName().$out;
-                    }
-                }
-            }
-        }
-
-        return false;
-    }
-
     public function editFieldAction(ContentType $contentType, FieldType $field, Request $request): Response
     {
         $editFieldType = new EditFieldType($field);
@@ -445,172 +392,6 @@ class ContentTypeController extends AbstractController
             'field' => $field,
             'contentType' => $contentType,
         ]);
-    }
-
-    /**
-     * Try to find (recursively) if there is a new field to add to the content type.
-     *
-     * @param array<mixed> $formArray
-     */
-    private function addNewSubfield(array $formArray, FieldType $fieldType): bool|string
-    {
-        if (\array_key_exists('subfield', $formArray)) {
-            if (isset($formArray['ems:internal:add:subfield:name'])
-                && 0 !== \strcmp((string) $formArray['ems:internal:add:subfield:name'], '')) {
-                if (static::isValidName($formArray['ems:internal:add:subfield:name'])) {
-                    $child = new FieldType();
-                    $child->setName($formArray['ems:internal:add:subfield:name']);
-                    $child->setType(SubfieldType::class);
-                    $child->setParent($fieldType);
-                    $fieldType->addChild($child);
-                    $this->logger->notice('log.contenttype.subfield.added', [
-                        'subfield_name' => $formArray['ems:internal:add:subfield:name'],
-                        EmsFields::LOG_OPERATION_FIELD => EmsFields::LOG_OPERATION_CREATE,
-                    ]);
-
-                    return '_ems_'.$child->getName().'_modal_options';
-                } else {
-                    $this->logger->error('log.contenttype.subfield.name_not_valid', [
-                        'field_format' => '/[a-z][a-z0-9_-]*/',
-                    ]);
-                }
-            } else {
-                $this->logger->error('log.contenttype.subfield.name_mandatory', [
-                ]);
-            }
-
-            return true;
-        } else {
-            /** @var FieldType $child */
-            foreach ($fieldType->getChildren() as $child) {
-                if (!$child->getDeleted()) {
-                    $out = $this->addNewSubfield($formArray['ems_'.$child->getName()], $child);
-                    if (false !== $out) {
-                        return '_ems_'.$child->getName().$out;
-                    }
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Try to find (recursively) if there is a field to duplicate.
-     *
-     * @param array<mixed> $formArray
-     */
-    private function duplicateField(array $formArray, FieldType $fieldType): bool|string
-    {
-        if (\array_key_exists('duplicate', $formArray)) {
-            if (isset($formArray['ems:internal:add:subfield:target_name'])
-                && 0 !== \strcmp((string) $formArray['ems:internal:add:subfield:target_name'], '')) {
-                if (static::isValidName($formArray['ems:internal:add:subfield:target_name'])) {
-                    $new = clone $fieldType;
-                    $new->setName($formArray['ems:internal:add:subfield:target_name']);
-                    if ($parent = $new->getParent()) {
-                        $parent->addChild($new);
-                    }
-
-                    $this->logger->notice('log.contenttype.field.added', [
-                        'field_name' => $formArray['ems:internal:add:subfield:target_name'],
-                        EmsFields::LOG_OPERATION_FIELD => EmsFields::LOG_OPERATION_CREATE,
-                    ]);
-
-                    return 'first_ems_'.$new->getName().'_modal_options';
-                } else {
-                    $this->logger->error('log.contenttype.field.name_not_valid', [
-                        'field_format' => '/[a-z][a-z0-9_-]*/ !'.Mapping::HASH_FIELD.' !'.Mapping::HASH_FIELD,
-                    ]);
-                }
-            } else {
-                $this->logger->error('log.contenttype.field.name_mandatory', [
-                ]);
-            }
-
-            return true;
-        } else {
-            /** @var FieldType $child */
-            foreach ($fieldType->getChildren() as $child) {
-                if (!$child->getDeleted()) {
-                    $out = $this->duplicateField($formArray['ems_'.$child->getName()], $child);
-                    if (false !== $out) {
-                        if (\is_string($out) && 'first' == \substr($out, 0, 5)) {
-                            return \substr($out, 5);
-                        }
-
-                        return '_ems_'.$child->getName().$out;
-                    }
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Try to find (recursively) if there is a field to remove from the content type.
-     *
-     * @param array<mixed> $formArray
-     */
-    private function removeField(array $formArray, FieldType $fieldType): bool
-    {
-        if (\array_key_exists('remove', $formArray)) {
-            $fieldType->setDeleted(true);
-            $this->logger->notice('log.contenttype.field.deleted', [
-                'field_name' => $fieldType->getName(),
-                EmsFields::LOG_OPERATION_FIELD => EmsFields::LOG_OPERATION_DELETE,
-            ]);
-
-            return true;
-        } else {
-            /** @var FieldType $child */
-            foreach ($fieldType->getChildren() as $child) {
-                if (!$child->getDeleted() && $this->removeField($formArray['ems_'.$child->getName()], $child)) {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Try to find (recursively) if there is a container where subfields must be reordered in the content type.
-     *
-     * @param array<mixed> $formArray
-     */
-    private function reorderFields(array $formArray, FieldType $fieldType): bool
-    {
-        if (\array_key_exists('reorder', $formArray)) {
-            /** @var string[] $keys */
-            $keys = \array_keys($formArray);
-            /** @var FieldType $child */
-            foreach ($fieldType->getChildren() as $child) {
-                if (!$child->getDeleted()) {
-                    $order = \array_search('ems_'.$child->getName(), $keys, true);
-                    if (false === $order || !\is_int($order)) {
-                        continue;
-                    }
-                    $child->setOrderKey($order);
-                }
-            }
-
-            $this->logger->notice('log.contenttype.field.reordered', [
-                'field_name' => $fieldType->getName(),
-            ]);
-
-            return true;
-        } else {
-            /** @var FieldType $child */
-            foreach ($fieldType->getChildren() as $child) {
-                if (!$child->getDeleted() && $this->reorderFields($formArray['ems_'.$child->getName()], $child)) {
-                    return true;
-                }
-            }
-        }
-
-        return false;
     }
 
     public function reorderAction(ContentType $contentType, Request $request): Response
@@ -783,51 +564,15 @@ class ContentTypeController extends AbstractController
                     'id' => $id,
                 ]);
             } else {
-                if ($out = $this->addNewField($inputContentType['fieldType'], $contentType->getFieldType())) {
-                    $contentType->getFieldType()->updateOrderKeys();
+                $openModal = $this->fieldTypeManager->handleRequest($contentType->getFieldType(), $inputContentType['fieldType']);
+                $contentType->getFieldType()->updateOrderKeys();
+                $em->persist($contentType);
+                $em->flush();
 
-                    $em->persist($contentType);
-                    $em->flush();
-
-                    return $this->redirectToRoute('contenttype.structure', [
-                        'id' => $id,
-                        'open' => $out,
-                    ]);
-                } elseif ($out = $this->addNewSubfield($inputContentType['fieldType'], $contentType->getFieldType())) {
-                    $contentType->getFieldType()->updateOrderKeys();
-                    $em->persist($contentType);
-                    $em->flush();
-
-                    return $this->redirectToRoute('contenttype.structure', [
-                        'id' => $id,
-                        'open' => $out,
-                    ]);
-                } elseif ($out = $this->duplicateField($inputContentType['fieldType'], $contentType->getFieldType())) {
-                    $contentType->getFieldType()->updateOrderKeys();
-                    $em->persist($contentType);
-                    $em->flush();
-
-                    return $this->redirectToRoute('contenttype.structure', [
-                        'id' => $id,
-                        'open' => $out,
-                    ]);
-                } elseif ($this->removeField($inputContentType['fieldType'], $contentType->getFieldType())) {
-                    $contentType->getFieldType()->updateOrderKeys();
-                    $em->persist($contentType);
-                    $em->flush();
-
-                    return $this->redirectToRoute('contenttype.structure', [
-                        'id' => $id,
-                    ]);
-                } elseif ($this->reorderFields($inputContentType['fieldType'], $contentType->getFieldType())) {
-                    // $contentType->getFieldType()->updateOrderKeys();
-                    $em->persist($contentType);
-                    $em->flush();
-
-                    return $this->redirectToRoute('contenttype.structure', [
-                        'id' => $id,
-                    ]);
-                }
+                return $this->redirectToRoute('contenttype.structure', \array_filter([
+                    'id' => $id,
+                    'open' => $openModal,
+                ]));
             }
         }
 
@@ -928,14 +673,5 @@ class ContentTypeController extends AbstractController
         $response->headers->set('Content-Disposition', $disposition);
 
         return $response;
-    }
-
-    protected function getDataFieldType(string $fieldTypeNameOrServiceName): DataFieldType
-    {
-        $dataFieldType = $this->formRegistry->getType($fieldTypeNameOrServiceName)->getInnerType();
-        if ($dataFieldType instanceof DataFieldType) {
-            return $dataFieldType;
-        }
-        throw new ElasticmsException(\sprintf('Expecting a DataFieldType instance, got a %s', $dataFieldType::class));
     }
 }
