+++ EMS/core-bundle/src/Resources/config/controllers.xml
@@ -83,7 +83,7 @@
             <argument type="service" id="emsco.logger" />
             <argument type="service" id="ems.service.contenttype" />
             <argument type="service" id="ems.service.mapping" />
-            <argument type="service" id="form.registry" />
+            <argument type="service" id="ems.form.field-type.manager" />
             <call method="setContainer"/>
             <tag name="container.service_subscriber"/>
             <tag name="controller.service_arguments"/>
